demo-app/requirements.txt
streamlit>=1.30.0
pyyaml>=6.0
jsonschema>=4.0
# Install AIGC from the parent directory:
#   pip install -e ../
# Or if aigc-sdk is published:
#   aigc-sdk>=0.3.0
